requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
fastapi==0.110.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.28.2
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.128.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-genai==1.52.0
google-generativeai==0.8.5
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.71.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.12.0
multidict==6.7.0
numpy==2.3.5
oauth2client==4.1.3
oauthlib==3.3.1
openai==2.8.1
openpyxl==3.1.5
packaging==25.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
tenacity==9.1.2
tqdm==4.67.1
twilio==9.8.7
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.29.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
